refactor: canvasrender目录结构更改
