+++ src/engine.ts
@@ -1,13 +1,13 @@
 import { EngineOption, Mdata, GridMdata } from './type';
 import { defaultEngineOption } from './config/engineoption';
-import { CanvasRender } from './view/render';
+import { CanvasRender } from './view';
 import { _merge } from './utils';
 import { Base } from './abstract/base';
 import { DataModel } from './model/mdata';
 export class Engine extends Base {
     // 数据
     private _sources: Mdata = null;
-    private canvasRender: CanvasRender;
+    private _canvasRender: CanvasRender;
     dataModel: DataModel;
     getDefaultCfg() {
         return {
@@ -16,27 +16,28 @@ export class Engine extends Base {
     }
     constructor(engineOpt: EngineOption) {
         super(engineOpt);
-        this.canvasRender = new CanvasRender(engineOpt.container);
+        this._canvasRender = new CanvasRender(engineOpt.container);
         // gridmap、srollindex等变量
         // 当这些变量更改的时候 需要让render定义去render
         this.dataModel = new DataModel({
             viewHeight: this.get('viewOption.viewHeight'),
             viewWidth: this.get('viewOption.viewWidth'),
-            canvasrender: this.canvasRender
+            canvasrender: this._canvasRender
         });
-        // this.rangeController = new RangeController();
-        // this.renderController = new RenderController(
-        //     this.rangeController
-        // );
-        // // datamodel 要调用render，所以将render作为data的子组件
-        // this.dataModel = new DataModel(
-        //     this.renderController, // renderControl里：rangecontroller
-        // );
         // // event要修改datamodel的值，event依赖datamodel
         // this.eventController = new EventController(
         //     this.dataModel
         // );
         // this.updateEngine();
+    }
+    getCommand() {
+        // 第一阶段：各个range的单独render独立开来
+        // 需要把每个range的依赖属性列出来，才能知道 外界设置什么可触发他render
+        // grid: 
+        // 第二阶段：将各个range的渲染挂在controller上调度
+        //         controller根据外界设置是x属性来判断如何调用(zindex)
+        // setRange().bgcolor = xxx;
+        // setRange().gridmap = xxx; // 如改动行高列宽
     }
     griddata(grid: GridMdata) {
         this.dataModel.resetGrid(grid);
