+++ src/view/index.ts
@@ -1,0 +1,1 @@
+export { CanvasRender } from './render/canvas';
